# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PharmaRoute - Biomedical Research Intelligence Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /search:
    post:
      operationId: search
      tags: [search]
      summary: Hybrid graph+vector search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchQuery"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"
  /search/voice:
    post:
      operationId: voiceSearch
      tags: [search]
      summary: Process voice query and return structured results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceQuery"
      responses:
        "200":
          description: Voice search result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated research reports
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a new research intelligence report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a specific report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
...
components:
  schemas:
...
    GraphStats:
      type: object
      required: [compounds, proteins, genes, diseases, pathways, papers, relationships]
      properties:
        compounds:
          type: integer
        proteins:
          type: integer
        genes:
          type: integer
        diseases:
          type: integer
        pathways:
          type: integer
        papers:
          type: integer
        relationships:
          type: integer
        lastUpdated:
          type: string
    GraphNode:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [Compound, Protein, Gene, Disease, Pathway, Paper]
        properties:
          type: object
          additionalProperties: true
        connections:
          type: integer
...
    PathResult:
      type: object
      required: [found, path]
      properties:
        found:
          type: boolean
        path:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        relationships:
          type: array
          items:
            type: string
        hops:
          type: integer
        confidence:
          type: number
...
    Discovery:
      type: object
      required: [id, title, compound, disease, confidence, impactScore, biologicalPath, evidenceCount]
      properties:
        id:
          type: string
        title:
          type: string
        compound:
          type: string
        disease:
          type: string
        confidence:
          type: number
        impactScore:
          type: number
        biologicalPath:
          type: array
          items:
            type: string
        evidenceCount:
          type: integer
        supportingPapers:
          type: array
          items:
            type: string
        explanation:
          type: string
        createdAt:
          type: string
...
    SearchQuery:
      type: object
      required: [query]
      properties:
        query:
          type: string
        nodeTypes:
          type: array
          items:
            type: string
        limit:
          type: integer
          default: 10
    VoiceQuery:
      type: object
      required: [transcript]
      properties:
        transcript:
          type: string
    SearchResult:
      type: object
      required: [query, nodes, discoveries, explanation]
      properties:
        query:
          type: string
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        discoveries:
          type: array
          items:
            $ref: "#/components/schemas/Discovery"
        explanation:
          type: string
        paths:
          type: array
          items:
            $ref: "#/components/schemas/PathResult"
    Report:
      type: object
      required: [id, title, type, createdAt, summary]
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [discovery, trends, graph_analysis, comprehensive]
        summary:
          type: string
        discoveries:
          type: array
          items:
            $ref: "#/components/schemas/Discovery"
        trends:
          type: array
          items:
            $ref: "#/components/schemas/ResearchTrend"
        graphStats:
          $ref: "#/components/schemas/GraphStats"
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [discovery, trends, graph_analysis, comprehensive]
        targetDisease:
          type: string
    ResearchTrend:
      type: object
      required: [topic, score, paperCount, direction]
      properties:
        topic:
          type: string
        score:
          type: number
        paperCount:
          type: integer
        direction:
          type: string
          enum: [rising, stable, declining]
        relatedTerms:
          type: array
          items:
            type: string
        description:
          type: string
...