# lib/api-zod/src/generated/types/graphStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaRoute - Biomedical Research Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */

export interface GraphStats {
  compounds: number;
  proteins: number;
  genes: number;
  diseases: number;
  pathways: number;
  papers: number;
  relationships: number;
  lastUpdated?: string;
}