# lib/api-zod/src/generated/types/researchTrend.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaRoute - Biomedical Research Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchTrendDirection } from './researchTrendDirection';

export interface ResearchTrend {
  topic: string;
  score: number;
  paperCount: number;
  direction: ResearchTrendDirection;
  relatedTerms?: string[];
  description?: string;
}